+++ backend/main.py
@@ -3,6 +3,8 @@ from app.modules.telemetry.mqtt_service import start_mqtt
 from contextlib import asynccontextmanager
 from app.modules.telemetry.router import router as telemetry_router
 from fastapi.middleware.cors import CORSMiddleware
+from app.core.database import create_db_and_tables
+
 
 
 
@@ -10,10 +12,13 @@ from fastapi.middleware.cors import CORSMiddleware
 @asynccontextmanager
 async def lifespan(app: FastAPI):
     # --- Lo que ocurre al ARRANCAR ---
+    print("Preparando base de datos...")
+    create_db_and_tables() # Para crear las tablas en la BBDD
     print("Iniciando conexión MQTT...")
     start_mqtt() 
     
-    yield # Aquí es donde la aplicación "vive"
+    
+    yield # Aquí es donde la app se queda funcionando
     
     # --- Lo que ocurre al CERRAR ---
     print("Cerrando conexiones...")
@@ -36,3 +41,5 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
+
+